Move NewCompletionCmd into client
