+++ server/util.go
@@ -233,38 +233,6 @@ func UpgradeOldPrivValFile(config *cfg.Config) {
 	}
 }
 
-// NewCompletionCmd builds a cobra.Command that generate bash completion
-// scripts for the given root command. If hidden is true, the command
-// will not show up in the root command's list of available commands.
-func NewCompletionCmd(rootCmd *cobra.Command, hidden bool) *cobra.Command {
-	flagZsh := "zsh"
-	cmd := &cobra.Command{
-		Use:   "completion",
-		Short: "Generate Bash/Zsh completion script to STDOUT",
-		Long: `To load completion script run
-
-. <(completion_script)
-
-To configure your bash shell to load completions for each session add to your bashrc
-
-# ~/.bashrc or ~/.profile
-. <(completion_script)
-`,
-		RunE: func(_ *cobra.Command, _ []string) error {
-			if viper.GetBool(flagZsh) {
-				return rootCmd.GenZshCompletion(os.Stdout)
-			}
-			return rootCmd.GenBashCompletion(os.Stdout)
-		},
-		Hidden: hidden,
-		Args:   cobra.NoArgs,
-	}
-
-	cmd.Flags().Bool(flagZsh, false, "Generate Zsh completion script")
-
-	return cmd
-}
-
 func skipInterface(iface net.Interface) bool {
 	if iface.Flags&net.FlagUp == 0 {
 		return true // interface down
